safari-600.7-branch/LayoutTests/svg/animations/animate-currentColor-expected.txt

SVG 1.1 dynamic animation tests

Tests animation on 'currentColor'.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS color.red.getFloatValue(CSSPrimitiveValue.CSS_NUMBER) is 255
PASS color.green.getFloatValue(CSSPrimitiveValue.CSS_NUMBER) is 0
PASS color.blue.getFloatValue(CSSPrimitiveValue.CSS_NUMBER) is 0
PASS color.red.getFloatValue(CSSPrimitiveValue.CSS_NUMBER) is 128
PASS color.green.getFloatValue(CSSPrimitiveValue.CSS_NUMBER) is 64
PASS color.blue.getFloatValue(CSSPrimitiveValue.CSS_NUMBER) is 0
PASS color.red.getFloatValue(CSSPrimitiveValue.CSS_NUMBER) is 0
PASS color.green.getFloatValue(CSSPrimitiveValue.CSS_NUMBER) is 128
PASS color.blue.getFloatValue(CSSPrimitiveValue.CSS_NUMBER) is 0
PASS successfullyParsed is true

TEST COMPLETE
